Move cache-related code to `shared` crate
