+++ crates/forge/tests/integration/setup_fork.rs
@@ -21,7 +21,6 @@ use crate::utils::test_case;
 use cheatnet::runtime_extensions::forge_runtime_extension::contracts_data::ContractsData;
 use forge::run_tests::package::RunForPackageArgs;
 use forge::shared_cache::FailedTestsCache;
-use forge_runner::DEFAULT_CACHE_DIR;
 use forge_runner::debugging::TraceArgs;
 use forge_runner::filtering::NameFilter;
 use forge_runner::forge_config::ForgeTrackedResource;
@@ -32,6 +31,7 @@ use forge_runner::running::target::prepare_test_target;
 use forge_runner::scarb::load_test_artifacts;
 use scarb_api::ScarbCommand;
 use scarb_api::metadata::metadata_for_dir;
+use shared::cache::DEFAULT_CACHE_DIR;
 use shared::test_utils::node_url::node_rpc_url;
 
 #[test]
