+++ crates/shared/Cargo.toml
@@ -9,6 +9,7 @@ workspace = true
 [dependencies]
 starknet_api.workspace = true
 anyhow.workspace = true
+camino.workspace = true
 starknet-types-core.workspace = true
 console.workspace = true
 semver.workspace = true
