+++ crates/shared/src/cache.rs
@@ -1,7 +1,12 @@
-use anyhow::{Context, Result};
+use anyhow::{Context, Result, bail, ensure};
+use camino::{Utf8Path, Utf8PathBuf};
+use std::env::{self, VarError};
 use std::fs::{self, File};
 use std::io::{ErrorKind, Write};
 use std::path::Path;
+
+pub const DEFAULT_CACHE_DIR: &str = ".snfoundry_cache";
+pub const USC_CACHE_DIR: &str = "universal-sierra-compiler";
 
 pub const CACHEDIR_TAG_FILENAME: &str = "CACHEDIR.TAG";
 pub const CACHEDIR_TAG_CONTENTS: &str = "\
@@ -10,6 +15,30 @@ Signature: 8a477f597d28d172789f06886806bc55
 # For information about cache directory tags, see:
 # https://bford.info/cachedir/
 ";
+
+pub fn resolve_cache_dir(workspace_root: &Utf8Path) -> Result<Utf8PathBuf> {
+    resolve_cache_dir_impl(workspace_root, env::var("SNFOUNDRY_CACHE"))
+}
+
+fn resolve_cache_dir_impl(
+    workspace_root: &Utf8Path,
+    cache_var: Result<String, VarError>,
+) -> Result<Utf8PathBuf> {
+    match cache_var {
+        Ok(cache_dir) => {
+            let cache_dir = Utf8PathBuf::from(cache_dir);
+            ensure!(
+                cache_dir.is_absolute(),
+                "SNFOUNDRY_CACHE must be an absolute path"
+            );
+            Ok(cache_dir)
+        }
+        Err(VarError::NotPresent) => Ok(workspace_root.join(DEFAULT_CACHE_DIR)),
+        Err(VarError::NotUnicode(_)) => {
+            bail!("SNFOUNDRY_CACHE must be a valid UTF-8 string")
+        }
+    }
+}
 
 pub fn prepare_cache_dir(cache_dir: impl AsRef<Path>) -> Result<()> {
     let cache_dir = cache_dir.as_ref();
@@ -41,7 +70,12 @@ pub fn prepare_cache_dir(cache_dir: impl AsRef<Path>) -> Result<()> {
 
 #[cfg(test)]
 mod tests {
-    use super::{CACHEDIR_TAG_CONTENTS, CACHEDIR_TAG_FILENAME, prepare_cache_dir};
+    use super::{
+        CACHEDIR_TAG_CONTENTS, CACHEDIR_TAG_FILENAME, DEFAULT_CACHE_DIR, prepare_cache_dir,
+        resolve_cache_dir_impl,
+    };
+    use camino::Utf8Path;
+    use std::env::VarError;
     use std::fs;
     use tempfile::TempDir;
 
@@ -54,6 +88,42 @@ mod tests {
 
         let tag = fs::read_to_string(cache_dir.join(CACHEDIR_TAG_FILENAME)).unwrap();
         assert_eq!(tag, CACHEDIR_TAG_CONTENTS);
+    }
+
+    #[test]
+    fn defaults_to_workspace_subdir_when_var_unset() {
+        let workspace = Utf8Path::new("/tmp/workspace");
+        assert_eq!(
+            resolve_cache_dir_impl(workspace, Err(VarError::NotPresent)).unwrap(),
+            workspace.join(DEFAULT_CACHE_DIR)
+        );
+    }
+
+    #[test]
+    fn accepts_absolute_custom_path() {
+        let resolved = resolve_cache_dir_impl(
+            Utf8Path::new("/tmp/workspace"),
+            Ok("/var/cache/snfoundry".to_string()),
+        )
+        .unwrap();
+        assert_eq!(resolved, Utf8Path::new("/var/cache/snfoundry"));
+    }
+
+    #[test]
+    fn rejects_relative_custom_path() {
+        let err = resolve_cache_dir_impl(
+            Utf8Path::new("/tmp/workspace"),
+            Ok("relative/cache".to_string()),
+        )
+        .unwrap_err();
+        assert_eq!(err.to_string(), "SNFOUNDRY_CACHE must be an absolute path");
+    }
+
+    #[test]
+    fn rejects_empty_string() {
+        let err =
+            resolve_cache_dir_impl(Utf8Path::new("/tmp/workspace"), Ok(String::new())).unwrap_err();
+        assert_eq!(err.to_string(), "SNFOUNDRY_CACHE must be an absolute path");
     }
 
     #[test]
